% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSamplr-package.R
\docType{package}
\name{dbSamplr}
\alias{dbSamplr}
\alias{dbSamplr-package}
\title{Database sampling tools}
\description{
Tools for designing database sampling schemes.  This can be done with
\code{gsDesign}, but this package is notably faster for large
databases.
}

