% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gsProbs}
\alias{gsProbs}
\title{Simulate simple multiple testing strategy for single binomial event}
\usage{
gsProbs(p, sizes, crits)
}
\arguments{
\item{p}{event probability}

\item{sizes}{monotonically increasing vector of sample sizes}

\item{crits}{vector below which sampling stops}
}
\value{
a vector giving probability of stopping at each sample
}
\description{
Simulate simple multiple testing strategy for single binomial event
}

